cos/queue_service.rs: colocate 12 tests with production (#984 P3 Phase 3a)
Move 12 tests from tx/mod.rs::tests to cos/queue_service.rs::tests
covering assign_local_dscp_rewrite, cos_batch_tx_made_progress,
drain_exact_local_fifo, estimate_cos_queue_wakeup_tick,
release_exact_local_scratch_frames, release_exact_prepared_scratch,
restore_cos_local_items, restore_cos_prepared_items,
settle_exact_local_fifo, settle_exact_prepared_fifo.

Adds: PROTO_TCP, COS_FLOW_FAIR_BUCKETS, CoSQueueConfig,
CoSQueueDropCounters, CoSQueueOwnerProfile, FlowRrRing.

Drops 7 stale cfg-test imports from tx/mod.rs that previously
pinned the moved fns (Codex round-1 cleanup).

Rebased onto current master after #1006 + #1004 merged.
cargo test --bins 865/0/2.
